src/requirements.txt
alembic==1.9.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Faker==15.3.4
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.45
sqlalchemy-seeder==0.3.0
tweepy==4.12.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
